Update layout naming and structure in `test.ui` and `test.py`
Changed the name of `horizontalLayout` to `upperLayout` in `test.ui` and `test.py` to reflect its actual position in the UI. In the same file, added `contentlayout` and `playerlayout` to `verticalLayout_2` which will be used for future UI enhancements. These changes will improve the clarity of layout positioning and make UI’s future modifications easier.