Updated UI style and implemented hot reloading
Updated the user interface in test.ui file to improve its appearance by adding new styles and enhanced PyQt builder with threading and time imports.

Removed the 'Build As Python File' button because file change detection in the background was implemented to auto-restart the application when changes are detected in either 'test.py' or 'test.ui'. Thus making the building process automatic and eliminating the need for manual build button.

Also, enhanced functionality by adding hot reloading mechanism using a separate thread. This enables the detection of changes in 'test.ui' and 'test.py' files so that the PyQt application can be restarted automatically when a change is detected.

Updated 'test.py' by adding more functionality to the PyQt application with QPushButtons, which have enriched functionalities with font style and icon amendments.