Refactor code for performance and readability
This commit primarily optimizes the reading of the 'app_config.json' file by only calling the json.load function once and storing the result in a variable. This reduces the I/O operation and improves code efficiency. In the Pyqtbuilder module, subprocesses are removed in favour of the os.system, simplifying the code for restarting the PyQt app. Layout adjustments in test.py enhance the UI look and feel. Also, unnecessary keys from playlists are removed, making data lightweight and ready for use. Other minor changes include updating button names and button icons to depend on the theme.